mysql-kafka/requirements.txt
mysql-connector-python>=8.0.32
kafka-python>=2.0.2
boto3>=1.26.137
pyarrow>=12.0.0
pandas>=1.5.3
PyYAML==6.0.1 
slack_sdk==3.19.0
requests>=2.28.2
psutil>=5.9.0
pyiceberg>=0.6.0
s3fs>=2023.12.0
duckdb>=0.9.0
sqlalchemy>=2.0.30
pyspark>=3.5.0
python-dotenv>=1.0.0